Added Employee and Admin Authorization and limited views for Employee

+++ RetailManagementTool.Models/Promotion/PromotionCreate.cs
@@ -19,6 +19,6 @@ namespace RetailManagementTool.Models.Promotion
         public int? PromoTypeId { get; set; }
 
         [Display(Name = "Promotion Value")]
-        public int PromotionValue { get; set; }
+        public decimal PromotionValue { get; set; }
     }
 }
